List only images compatible with the box in use (read opkg/arch.conf for box without boxbranding) like in the plugin
This would allow you to swap the usb key between boxes and choose the correct image.

Before swapping remember to switch to the image in flash memory.

Please note that this fix isn't softbrick safe yet.

We need to differentiate the kernel backup per box by renaming it in flash-boxname.bin